+++ panda/src/linmath/lvecBase2_src.h
@@ -44,7 +44,7 @@ PUBLISHED:
 #ifdef HAVE_PYTHON
   INLINE_LINMATH void __setitem__(int i, FLOATTYPE v);
 #endif
-  INLINE_LINMATH static size_t size();
+  INLINE_LINMATH static int size();
 
   INLINE_LINMATH bool is_nan() const;
